Add test fixtures and reduce redundant code
